Increasing the log rendering performance & showing icons in level selector
